fix(circular_queue): pass z as int* to FUN_00114fa0 (inflate_fast)
The zlib upstream-source lift (ec56230c) retyped FUN_00114fa0's param_6
from `int` to `int *` in kb.json and updated the definition, but not the
call site in inflate_codes, so the TU stopped compiling:

  circular_queue.c:201: incompatible integer to pointer conversion
  passing 'int' to parameter of type 'int *'

`z` is already declared `int *z` in inflate_codes and param_6 is the
zlib z_streamp, so the `(int)` downcast was stale, not load-bearing.
Dropping it restores the build with no change in the value passed.

This broke `main` itself, not just this branch: the .c file and the
generated decl are byte-identical on both, so the full build (and with
it every auto-session land gate) failed on main until now.

+++ src/halo/memory/circular_queue.c
@@ -198,7 +198,7 @@ int inflate_codes(unsigned int s, int *z, int r)
         UPDATE
         r = FUN_00114fa0((int)*((unsigned char *)c + 0x10),
                          (int)*((unsigned char *)c + 0x11), (int)c[5],
-                         (int)c[6], (int)s, (int)z);
+                         (int)c[6], (int)s, z);
         LOAD
         if (r != 0) {
           *c = r == 1 ? 7 : 9;
